Drop BigInteger from the crack time estimate
The guess count only needed arbitrary precision because it was built as an
integer 2^bits. The display shows nothing but the magnitude, so the estimate
is now carried as a base-10 exponent in a double and converted once at the
end. Hash rates move from BigInteger to ulong; the largest profile is 10^15.

Flooring the exponent is gone with it, so fractional entropy values no longer
report a crack time up to twice too low.

+++ MainForm.cs
@@ -616,7 +616,7 @@ public partial class MainForm : Form
         var profiles = PasswordGenerator.ThreatProfiles;
         for (int i = profiles.Length - 1; i >= 0; i--)
         {
-            if (hashRate >= (ulong)profiles[i].HashesPerSecond)
+            if (hashRate >= profiles[i].HashesPerSecond)
             {
                 return profiles[i];
             }
